requirements.txt
python-dotenv
psycopg
psycopg2-binary
requests
pytelegrambotapi
schedule